requirements.txt
altair==4.2.2
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
entrypoints==0.4
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.2
numpy==1.26.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
py4j==0.10.9.7
pyarrow==18.1.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jobspy==1.1.75
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.40.2
streamlit-aggrid==1.0.5
streamlit-awesome-table==0.1.0
tenacity==9.0.0
tls-client==1.0.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
duckdb==1.2.0